lambda_functions/createImagesFromPdf/requirements/layer.txt
pdf2image==1.9.0
awscli==1.16.231
botocore==1.12.221
six~=1.11.0
boto3==1.9.221
pillow==6.2.0
opencv-python==4.1.1.26